[android] Code refactored for OsmUploadWork
Unnecessary classes like JobIdMap.java and MwmWork.java removed to simplify the code, added internet constrain to OsmUploadWork, added if OSM changes available check before starting worker #4531
